requirements.txt
# ML Yacht Autopilot - Dependencies
# ==================================

# Core
numpy>=1.21.0

# CAN Bus
python-can>=4.0.0

# Serial communication  
pyserial>=3.5

# ML - TensorFlow (choose one based on platform)
# For development/training (full TensorFlow):
tensorflow>=2.10.0

# For Raspberry Pi inference only (lightweight):
# tflite-runtime>=2.10.0

# Hardware interfaces (Raspberry Pi only)
# RPi.GPIO
# adafruit-circuitpython-ads1x15

# Testing and development
pytest>=7.0.0
pytest-cov>=3.0.0

# Optional: visualization
# matplotlib>=3.5.0
